requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
click==8.2.1
colorama==0.4.6
cryptography>=41.0.0
Django>=4.2,<5.0
django-allauth>=0.60.0
djangorestframework>=3.14.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.4.1
opencv-python==4.13.0.90
orjson==3.11.7
packaging==26.0
pefile==2024.8.26
Pillow>=10.0.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
python-dotenv==1.2.1
python-multipart==0.0.22
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests>=2.25.0
rich==14.3.2
rich-toolkit==0.18.1
rignore==0.7.6
sentry-sdk==2.51.0
setuptools==80.10.1
shellingham==1.5.4
SQLAlchemy==2.0.41
starlette==0.50.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.14.1
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.3
